"update dependencies and clean phpunit config
This commit refactors minor portions of the code, with a focus on cleanup and simplification. In addition, obsolete and redundant PHPUnit configurations have been deleted, and several dependencies in composer.lock have been updated to their latest versions to leverage new features and ensure potential bugs or security vulnerabilities are patched. The PHPUnit XML Schema version also changed from 9.5 to 10.3 to match with the PHPUnit version used. Lastly, 'phpunit.cache' location is updated for better caching."